🐛 Fix portfolio & dashboard OG
