+++ src/locales/en.json
@@ -26,6 +26,8 @@
   "campaign_nft_upcoming": "Next Writing NFT coming on",
   "campaign_nft_view_nft_dashboard": "Visit NFT Dashboard",
   "clipboard_copied": "Copied",
+  "dashboard_title": "My Dashbaord",
+  "dashboard_description": "Connect your wallet to view your NFT dashboard.",
   "dashboard_portfolio_button": "Public Portfolio",
   "faq_list": [
     {
